% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_ais_hex.R
\name{decode_ais_hex}
\alias{decode_ais_hex}
\title{Decode AIS hex string to R list}
\usage{
decode_ais_hex(hex_string)
}
\arguments{
\item{hex_string}{A character string containing the AIS message in hex format}
}
\value{
A list of the AIS message
}
\description{
Decode AIS hex string to R list
}
