% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket_client.R
\name{connect_ais_stream}
\alias{connect_ais_stream}
\title{Receive real-time vessel AIS messages.}
\usage{
connect_ais_stream(
  api_key,
  bounding_box = NULL,
  file_path = NULL,
  layer = NULL,
  verbose = TRUE,
  reconnect_delay = 5,
  heartbeat_interval = NULL,
  outDir = NULL
)
}
\arguments{
\item{api_key}{Your AISStream API key as a string.}

\item{bounding_box}{List of coordinates: list(list(c(lat1, lon1), c(lat2, lon2))).}

\item{file_path}{Character string to a shapefile, KML, or KMZ file.}

\item{layer}{Character string specifying the layer name in the file if it contains multiple layers.}

\item{verbose}{Logical. If TRUE, status messages will be printed to the console (default: TRUE).}

\item{reconnect_delay}{Seconds to wait before reconnecting after a disconnect (default: 5)}

\item{heartbeat_interval}{Seconds between optional heartbeat messages (default: NULL, no heartbeat)}

\item{outDir}{Character string specifying the output directory to save data. If NULL, data is not saved.}
}
\value{
A WebSocket object
}
\description{
Connect through the AISStream.io WebSocket to receive real-time vessel AIS messages for a specified bounding box or spatial file.
}
